pyproject.toml
[project]
name = "file-organizer-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.8.0",
    "pillow>=11.2.1",
]
